Added suppression and tests

+++ binaryaudit/dnf.py
@@ -2,12 +2,13 @@ from binaryaudit import run
 from binaryaudit import util
 
 import json
-import rpmfile
 import os
+import rpmfile
+import shutil
 import subprocess
 import urllib.request
 
-def process_downloads(source_dir, new_json_file, old_json_file, output_dir):
+def process_downloads(source_dir, new_json_file, old_json_file, output_dir, cleanup=True):
     ''' Finds and downloads older versions of RPMs.
 
         Parameters:
@@ -16,24 +17,30 @@ def process_downloads(source_dir, new_json_file, old_json_file, output_dir):
             old_json_file (str): The name of the JSON file containing the older set of packages
             output_dir (str): The path to the output directory of abipkgdiff
     '''
-    os.mkdir(source_dir + "old")
-    old_rpm_dict = {}
-    with open(new_json_file, "r") as file:
-        data = json.load(file)
-    for key, values in data.items():
-        for value in values:
-            with rpmfile.open(source_dir + value) as rpm:
-                name = rpm.headers.get("name")            
-            old_rpm_name = download(key, source_dir, name, old_rpm_dict)
-        if old_rpm_name == "":
-            continue
-        with open(old_json_file, "w") as outputFile:
-            json.dump(old_rpm_dict, outputFile, indent=2)
-        generate_abidiffs(key, source_dir, new_json_file, old_json_file, output_dir)
-    os.rmdir(source_dir + "old")
-    os.remove(new_json_file)
-    os.remove(old_json_file)
-
+    try:
+        if not os.path.exists(source_dir + "old"):
+            os.mkdir(source_dir + "old")
+        old_rpm_dict = {}
+        with open(new_json_file, "r") as file:
+            data = json.load(file)
+        for key, values in data.items():
+            for value in values:
+                with rpmfile.open(source_dir + value) as rpm:
+                    name = rpm.headers.get("name") 
+                old_rpm_name = download(key, source_dir, name, old_rpm_dict)
+            if old_rpm_name == "":
+                continue
+            with open(old_json_file, "w") as outputFile:
+                json.dump(old_rpm_dict, outputFile, indent=2)
+            generate_abidiffs(key, source_dir, new_json_file, old_json_file, output_dir, cleanup)
+    finally:
+        if cleanup == True:
+            try:
+                shutil.rmtree(source_dir + "old")
+                os.remove(new_json_file)
+                os.remove(old_json_file)
+            except:
+                pass
 def download(key, source_dir, name, old_rpm_dict):
     ''' Finds and downloads older versions of RPMs.
 
@@ -43,19 +50,19 @@ def download(key, source_dir, name, old_rpm_dict):
             name: The name of the RPM
             old_rpm_dict: The dictionary containing the older set of packages
     '''
-    docker, docker_exit_code = run.run_command(["sudo", "docker", "run", "-t", "--rm", "mariner:abidiff", "/usr/bin/dnf", "repoquery", "--quiet", "--latest-limit=1", name], None, subprocess.PIPE)
+    docker, docker_exit_code = run.run_command_docker(["/usr/bin/dnf", "repoquery", "--quiet", "--latest-limit=1", name], None, subprocess.PIPE)
     old_rpm_name = docker.stdout.read().decode('utf-8')
     if old_rpm_name == "":
         return old_rpm_name
     old_rpm_name = old_rpm_name.replace("\r\n", "")
-    docker_loc, docker_loc_exit_code = run.run_command(["sudo", "docker", "run", "-t", "--rm", "mariner:abidiff", "/usr/bin/dnf", "repoquery", "--quiet", "--location", "--latest-limit=1", name], None, subprocess.PIPE)
+    docker_loc, docker_loc_exit_code = run.run_command_docker(["/usr/bin/dnf", "repoquery", "--quiet",  "--location", "--latest-limit=1", name], None, subprocess.PIPE)
     url = docker_loc.stdout.read().decode('utf-8')
     util.debug("url: {}".format(url))
-    urllib.request.urlretrieve(url, source_dir + "/old/" + old_rpm_name)
+    urllib.request.urlretrieve(url, source_dir + "old/" + old_rpm_name)
     old_rpm_dict.setdefault(key, []).append(old_rpm_name)
     return old_rpm_name
 
-def generate_abidiffs(key, source_dir, new_json_file, old_json_file, output_dir):
+def generate_abidiffs(key, source_dir, new_json_file, old_json_file, output_dir, cleanup=True):
     ''' Runs abipkgdiff against the grouped packages.
 
         Parameters:
@@ -97,21 +104,29 @@ def generate_abidiffs(key, source_dir, new_json_file, old_json_file, output_dir)
     i = 0
     for rpm in rpms_with_so:
         if i%2 == 0:
-            command_list = ["abipkgdiff"]
+            command_list = ["abipkgdiff", "--suppressions", "../conf/suppressions.conf"]
+            old_main_rpm = rpm
         i += 1
         command_list.append(rpm)
         if i%2 == 1:
             continue
+        new_main_rpm = rpm
         for arg in cmd_supporting_args:
             command_list.append(arg)
         with open("output_file", "w") as output_file:
             abipkgdiff, abipkgdiff_exit_code = run.run_command(command_list, None, output_file)
             if abipkgdiff_exit_code != 0:
                 with rpmfile.open(old_main_rpm) as rpm:
-                    fileName = rpm.headers.get('name').decode('utf-8') + "__" + rpm.headers.get('version').decode('utf-8') + "-" + rpm.headers.get('release').decode('utf-8') + "__"
+                    name = rpm.headers.get('name').decode('utf-8')
+                    old_version = rpm.headers.get('version').decode('utf-8')
+                    old_release = rpm.headers.get('release').decode('utf-8')
                 with rpmfile.open(new_main_rpm) as rpm:
-                    fileName += rpm.headers.get('version').decode('utf-8') + "-" + rpm.headers.get('release').decode('utf-8') + ".abidiff"
+                    new_version = rpm.headers.get('version').decode('utf-8')
+                    new_release = rpm.headers.get('release').decode('utf-8')
+                print("Incompatibility found between " + name + "-" + old_version + "-" + old_release + " and " + name + "-" + new_version + "-" + new_release)
+                fileName = name + "__" + old_version + "-" + old_release + "__" + new_version + "-" + new_release + ".abidiff"
                 os.rename("output_file", output_dir + fileName)
-    for value in old_data[key]:
-        os.remove(source_dir + "old/" + value)
+    if cleanup == True:
+        for value in old_data[key]:
+            os.remove(source_dir + "old/" + value)
     output_file.close()
